Add picard.util.qt and move ReadWriteLockContext into it
Introduce picard.util.qt as the home for utility helpers that depend on
PyQt6. Keeping them separate from picard.util means importing picard.util
does not pull in Qt, so the bulk of the helpers stay usable from Qt-free
contexts (build tooling, potential headless/alternative frontends).

Start by moving ReadWriteLockContext (which wraps QReadWriteLock) into it.
Update the importer (metadata). Subsequent commits move the remaining
Qt-dependent helpers.
